Reorganise headers by hash family
