--- Sming/Components/crypto/include/Crypto.h
@@ -1,13 +1,0 @@
-/****
- * Sming Framework Project - Open Source framework for high efficiency native ESP8266 development.
- * Created 2015 by Skurydin Alexey
- * http://github.com/SmingHub/Sming
- * All files of the Sming Core are provided under the LGPL v3 license.
- *
- * Crypto.h
- *
- ****/
-
-#pragma once
-
-#include "Crypto/HmacContext.h"
