+++ Sming/Components/crypto/include/Crypto/HashContext.h
@@ -10,8 +10,6 @@
 
 #pragma once
 
-#include "HashEngine.h"
-#include "Blake2s.h"
 #include "Blob.h"
 #include "ByteArray.h"
 
@@ -131,20 +129,5 @@ public:
 private:
 	Engine engine;
 };
-
-/*
- * Context definitions
- */
-
-using Md5 = HashContext<Md5Engine>;
-using Sha1 = HashContext<Sha1Engine>;
-using Sha224 = HashContext<Sha224Engine>;
-using Sha256 = HashContext<Sha256Engine>;
-using Sha384 = HashContext<Sha384Engine>;
-using Sha512 = HashContext<Sha512Engine>;
-
-template <size_t hashsize> using Blake2s = HashContext<Blake2sEngine<hashsize>>;
-using Blake2s256 = Blake2s<32>;
-using Blake2s128 = Blake2s<16>;
 
 } // namespace Crypto
