+++ Sming/Components/crypto/include/Crypto/HashEngine.h
@@ -4,20 +4,25 @@
  * http://github.com/SmingHub/Sming
  * All files of the Sming Core are provided under the LGPL v3 license.
  *
- * HashEngine.h
+ * HashEngine.h - Hash engine template code
  *
  ****/
 
 #pragma once
 
-#include "HashApi.h"
-
-namespace Crypto
-{
 /**
- * @brief Hash engines just call the C API so class can be constructed using a macro
+ * @brief Macro template to construct Engine class wrapper for a given hash API
+ * @param class_ Name for the class type, upper-case camel
+ * @param name_ Name for the C API, lower-case camel
+ * @param hashsize_ Size of hash in bytes
+ * @param statesize_ Size of state in bytes
+ * @param blocksize_ Size of block buffer in bytes
+ * @param INIT Macro to implement init/state and any other custom interface code
+ *
+ * Init methods may have optional parameters.
+ * State methods are only required for standard hashes. These are used by Bear SSL multi-hash implementation.
  */
-#define CRYPTO_HASH_ENGINE(class_, name_, hashsize_, statesize_, blocksize_)                                           \
+#define CRYPTO_HASH_ENGINE(class_, name_, hashsize_, statesize_, blocksize_, INIT)                                     \
 	class class_##Engine                                                                                               \
 	{                                                                                                                  \
 	public:                                                                                                            \
@@ -26,10 +31,7 @@ namespace Crypto
 		static constexpr size_t statesize = statesize_;                                                                \
 		static constexpr size_t blocksize = blocksize_;                                                                \
                                                                                                                        \
-		void init()                                                                                                    \
-		{                                                                                                              \
-			CRYPTO_NAME(name_, init)(&ctx);                                                                            \
-		}                                                                                                              \
+		INIT(name_)                                                                                                    \
                                                                                                                        \
 		void update(const void* data, size_t size)                                                                     \
 		{                                                                                                              \
@@ -41,29 +43,31 @@ namespace Crypto
 			CRYPTO_NAME(name_, final)(hash, &ctx);                                                                     \
 		}                                                                                                              \
                                                                                                                        \
-		uint64_t get_state(void* state)                                                                                \
-		{                                                                                                              \
-			return CRYPTO_NAME(name_, get_state)(&ctx, state);                                                         \
-		}                                                                                                              \
-                                                                                                                       \
-		void set_state(const void* state, uint64_t count)                                                              \
-		{                                                                                                              \
-			CRYPTO_NAME(name_, set_state)(&ctx, state, count);                                                         \
-		}                                                                                                              \
-                                                                                                                       \
 	private:                                                                                                           \
 		CRYPTO_CTX(name_) ctx;                                                                                         \
 	};
 
-/*
- * Engine definitions
+/**
+ * @brief Macro template to construct standard hash engines
  */
+#define CRYPTO_HASH_ENGINE_STD(class_, name_, hashsize_, statesize_, blocksize_)                                       \
+	CRYPTO_HASH_ENGINE(class_, name_, hashsize_, statesize_, blocksize_, CRYPTO_HASH_ENGINE_STD_INIT)
 
-CRYPTO_HASH_ENGINE(Md5, md5, MD5_SIZE, MD5_STATESIZE, MD5_BLOCKSIZE);
-CRYPTO_HASH_ENGINE(Sha1, sha1, SHA1_SIZE, SHA1_STATESIZE, SHA1_BLOCKSIZE);
-CRYPTO_HASH_ENGINE(Sha224, sha224, SHA224_SIZE, SHA224_STATESIZE, SHA224_BLOCKSIZE);
-CRYPTO_HASH_ENGINE(Sha256, sha256, SHA256_SIZE, SHA256_STATESIZE, SHA256_BLOCKSIZE);
-CRYPTO_HASH_ENGINE(Sha384, sha384, SHA384_SIZE, SHA384_STATESIZE, SHA384_BLOCKSIZE);
-CRYPTO_HASH_ENGINE(Sha512, sha512, SHA512_SIZE, SHA512_STATESIZE, SHA512_BLOCKSIZE);
-
-} // namespace Crypto
+/**
+ * @brief Macro template to provide init/state methods for standard hash engines
+ */
+#define CRYPTO_HASH_ENGINE_STD_INIT(name_)                                                                             \
+	void init()                                                                                                        \
+	{                                                                                                                  \
+		CRYPTO_NAME(name_, init)(&ctx);                                                                                \
+	}                                                                                                                  \
+                                                                                                                       \
+	uint64_t get_state(void* state)                                                                                    \
+	{                                                                                                                  \
+		return CRYPTO_NAME(name_, get_state)(&ctx, state);                                                             \
+	}                                                                                                                  \
+                                                                                                                       \
+	void set_state(const void* state, uint64_t count)                                                                  \
+	{                                                                                                                  \
+		CRYPTO_NAME(name_, set_state)(&ctx, state, count);                                                             \
+	}
