+++ Sming/Components/crypto/src/sha2.h
@@ -19,7 +19,8 @@
 
 #pragma once
 
-#include "hash.h"
+#include "stdhash.h"
+#include "../include/Crypto/HashApi/sha2.h"
 
 namespace Crypto
 {
